:truck: Move a part of @cuillere/server into @cuillere/server-plugin
