+++ crud/package.json
@@ -47,7 +47,7 @@
     "ci": "yarn build:node-version && yarn build:types && yarn lint && yarn test"
   },
   "devDependencies": {
-    "@cuillere/server": "^0.13.0"
+    "@cuillere/server-plugin": "^0.13.0"
   },
   "peerDependencies": {
     "@cuillere/core": "^0.13.0"
