+++ postgres/src/server-plugin.ts
@@ -1,5 +1,5 @@
 import { cuillere } from '@cuillere/core'
-import { ServerContext, ServerPlugin, taskManagerPlugin } from '@cuillere/server'
+import { ServerContext, ServerPlugin, taskManagerPlugin } from '@cuillere/server-plugin'
 import { registerCrudProvider } from '@cuillere/crud'
 
 import { getClientManager } from './client-manager'
@@ -24,9 +24,6 @@ export function postgresServerPlugin(config: PostgresConfig) {
     registerCrudProvider(srvCtx, 'postgres', {
       build() {
         return cuillere(
-          // FIXME using taskManagerPlugin means cuillere/server is not a devDependency anymore
-          // either move postgresServerPlugin in @cuillere/server-postgres
-          // or move a part of @cuillere/server in a third package
           taskManagerPlugin(
             getClientManager({
               poolManager,
