+++ server-plugin/tsconfig-node12.json
@@ -1,0 +1,10 @@
+{
+  "extends": "../tsconfig-node12.json",
+  "compilerOptions": {
+    "outDir": "dist/node/12",
+    "rootDir": "src"
+  },
+  "exclude": [
+    "src/**/*.spec.ts"
+  ]
+}
