+++ server-plugin/tsconfig-node14.json
@@ -1,0 +1,10 @@
+{
+  "extends": "../tsconfig-node14.json",
+  "compilerOptions": {
+    "outDir": "dist/node/14",
+    "rootDir": "src"
+  },
+  "exclude": [
+    "src/**/*.spec.ts"
+  ]
+}
