+++ server/src/config.ts
@@ -1,0 +1,6 @@
+import { ServerContext, ServerPlugin } from '@cuillere/server-plugin'
+
+export interface CuillereConfig {
+  contextKey?: string
+  plugins?: ((srvCtx: ServerContext) => ServerPlugin)[]
+}
