+++ server/src/index.ts
@@ -1,10 +1,6 @@
 export * from './apollo-server-plugin'
+export * from './config'
 export * from './exports'
 export * from './koa-middleware'
 export { makeExecutableSchema } from './schema'
-export * from './server-plugin'
 export * from './server'
-export * from './task-executor'
-export * from './task-manager'
-export * from './transaction-manager'
-export * from './types'
